Move messages package to into sequencediagrams
