+++ src/uml/diagrams/sequence/sequencediagrams/messages/MessageAsync.java
@@ -1,19 +1,19 @@
-package uml.diagrams.sequence.messages;
+package uml.diagrams.sequence.sequencediagrams.messages;
 
 import uml.diagrams.sequence.exceptions.MessageFormatException;
 import uml.diagrams.sequence.exceptions.SequenceDiagramRuleException;
 import uml.diagrams.sequence.lifelines.Lifeline;
 
-public class Reply extends Message {
+public class MessageAsync extends Message {
 
-    public Reply(String name, Float prob, Lifeline source, Lifeline target)
+    public MessageAsync(String name, Float prob, Lifeline source, Lifeline target)
             throws MessageFormatException, SequenceDiagramRuleException {
         super(name, prob, source, target);
     }
     
     @Override
     public String toString() {
-        return "<Reply " + 
+        return "<MessageAsync " + 
                 "name=\"" + this.getName() + "\" " + 
                 "prob=\"" + this.prob + "\" " + 
                 "source=\"" + this.getSource().getName() + "\" " +
