+++ src/uml/diagrams/sequence/sequencediagrams/messages/Reply.java
@@ -1,19 +1,19 @@
-package uml.diagrams.sequence.messages;
+package uml.diagrams.sequence.sequencediagrams.messages;
 
 import uml.diagrams.sequence.exceptions.MessageFormatException;
 import uml.diagrams.sequence.exceptions.SequenceDiagramRuleException;
 import uml.diagrams.sequence.lifelines.Lifeline;
 
-public class MessageAsync extends Message {
+public class Reply extends Message {
 
-    public MessageAsync(String name, Float prob, Lifeline source, Lifeline target)
+    public Reply(String name, Float prob, Lifeline source, Lifeline target)
             throws MessageFormatException, SequenceDiagramRuleException {
         super(name, prob, source, target);
     }
     
     @Override
     public String toString() {
-        return "<MessageAsync " + 
+        return "<Reply " + 
                 "name=\"" + this.getName() + "\" " + 
                 "prob=\"" + this.prob + "\" " + 
                 "source=\"" + this.getSource().getName() + "\" " +
