+++ test/uml/diagrams/sequence/sequencediagrams/SequenceDiagramCreateElementsTest.java
@@ -12,11 +12,10 @@ import org.junit.jupiter.params.provider.MethodSource;
 import uml.diagrams.sequence.exceptions.MessageFormatException;
 import uml.diagrams.sequence.exceptions.SequenceDiagramRuleException;
 import uml.diagrams.sequence.lifelines.Lifeline;
-import uml.diagrams.sequence.messages.Message;
-import uml.diagrams.sequence.messages.MessageAsync;
-import uml.diagrams.sequence.messages.MessageSync;
-import uml.diagrams.sequence.messages.Reply;
-import uml.diagrams.sequence.sequencediagrams.SequenceDiagram;
+import uml.diagrams.sequence.sequencediagrams.messages.Message;
+import uml.diagrams.sequence.sequencediagrams.messages.MessageAsync;
+import uml.diagrams.sequence.sequencediagrams.messages.MessageSync;
+import uml.diagrams.sequence.sequencediagrams.messages.Reply;
 
 public class SequenceDiagramCreateElementsTest {
     public static Collection<Object[]> sequenceDiagrams()
